simplified etcd setup
to keep @pires happy there there is now a single-node etcd cluster on master
and the minions have local etcds running in proxy/client mode pointing to master.

on my personal roadmap is to extend the tooling to just consume pragmatically
either one of ...
- an external etcd discovery server (public, private, whatever)
- an external etcd cluster or server
- a single, or multi-node local etcd cluster

anyway right now, it's back to basics.

Signed-off-by: António Meireles <antonio.meireles@reformi.st>