do not hard-code flannel to eth1
by default flannel will default to *the interface for the default route on the
machine* (sic) which may or not be eth1 (isn't e.g. in VMware Fusion).

One way to achieve maximum portability is simply not *tell* anything to flannel
on cloud-config by way of which it will just abide by its' defaults, another,
more verbose, that we do here, functionally equivalent, is just to tell him
to use the VM's public interface address.

cf. pires/kubernetes-vagrant-coreos-cluster#76
cf. pires/kubernetes-vagrant-coreos-cluster#82

Signed-off-by: António Meireles <antonio.meireles@reformi.st>